+++ NEWS.md
@@ -3,6 +3,9 @@
 ### Enhancements
 * Added `numeric_formats` argument to `tm_t_summary` to customize formatting of numeric statistics.
 
+### Miscellaneous
+* Multiple decorators can be applied to the same output object (#1465).
+
 # teal.modules.clinical 0.12.0
 
 ### Enhancements